requirements.txt
slack-bolt
PyPDF2
Pillow
pytesseract
pdf2image
langchain
requests
openai
langchain-community
chromadb
tiktoken
